+++ ci/licenses_golden/excluded_files
@@ -124,6 +124,7 @@
 ../../../flutter/impeller/.clang-format
 ../../../flutter/impeller/.gitignore
 ../../../flutter/impeller/README.md
+../../../flutter/impeller/aiks/aiks_blend_unittests.cc
 ../../../flutter/impeller/aiks/aiks_blur_unittests.cc
 ../../../flutter/impeller/aiks/aiks_gradient_unittests.cc
 ../../../flutter/impeller/aiks/aiks_path_unittests.cc
