+++ impeller/aiks/BUILD.gn
@@ -85,6 +85,7 @@ impeller_component("context_spy") {
 template("aiks_unittests_component") {
   target_name = invoker.target_name
   predefined_sources = [
+    "aiks_blend_unittests.cc",
     "aiks_blur_unittests.cc",
     "aiks_gradient_unittests.cc",
     "aiks_path_unittests.cc",
